Deleted Models
This was due to maximum WEBGL Support pages, worked on the contact page, finished the functionality. Need to connect with a Web Server/Host + Will integrate further security and utilities such as limiting characters and accepting only official emails. Will need to design the Background and add Personal Projects later.